+++ pom.xml
@@ -5,7 +5,7 @@
   <modelVersion>4.0.0</modelVersion>
 
   <groupId>edu.akka.sample</groupId>
-  <artifactId>finance</artifactId>
+  <artifactId>app</artifactId>
   <version>1.0-SNAPSHOT</version>
 
   <repositories>
