+++ src/test/java/edu/akka/sample/app/data/provider/TestTransactionProvider.java
@@ -22,13 +22,13 @@
  * SOFTWARE.
  */
 
-package edu.akka.sample.finance.data.provider;
+package edu.akka.sample.app.data.provider;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import edu.akka.sample.finance.data.definition.Transaction;
+import edu.akka.sample.app.data.definition.Transaction;
 import java.util.List;
 import org.junit.jupiter.api.Test;
